Move simulate.py alongside other minefield widgets
